Massive amounts of refactoring to move modules to where they should be. Backwards compatible, but expect lots of Deprecation messages.
