+++ pymatgen/analysis/structure_prediction/data/lambda.json
