+++ pymatgen/analysis/structure_prediction/tests/test_substitutor.py
@@ -10,7 +10,7 @@ import json
 
 from pymatgen.core.periodic_table import Specie
 from pymatgen.core.composition import Composition
-from pymatgen.structure_prediction.substitutor import Substitutor
+from pymatgen.analysis.structure_prediction.substitutor import Substitutor
 
 
 def get_table():
@@ -19,7 +19,7 @@ def get_table():
     initialization time, and make unit tests insensitive to changes in the
     default lambda table.
     """
-    data_dir = os.path.join(os.path.dirname(__file__), "..", "..", "..",
+    data_dir = os.path.join(os.path.dirname(__file__), "..", "..", "..", "..",
                             'test_files', "struct_predictor")
     json_file = os.path.join(data_dir, 'test_lambda.json')
     with open(json_file) as f:
