--- pymatgen/matproj/tests/__init__.py
@@ -1,11 +1,0 @@
-"""
-TODO: Modify module doc.
-"""
-
-__author__ = "Shyue Ping Ong"
-__copyright__ = "Copyright 2012, The Materials Project"
-__version__ = "0.1"
-__maintainer__ = "Shyue Ping Ong"
-__email__ = "shyuep@gmail.com"
-__date__ = "2/12/13"
-
