+++ pymatgen/serializers/json_coders.py
@@ -1,47 +1,9 @@
-# coding: utf-8
-# Copyright (c) Pymatgen Development Team.
-# Distributed under the terms of the MIT License.
+import warnings
 
-from __future__ import division, unicode_literals
+warnings.warn("pymatgen.serializers.json_coders has been moved to "
+              "pymatgen.util.serialization. This stub will be removed in pmg "
+              "2018.01.01.",
+              DeprecationWarning)
 
-"""
-Most features of this module has been moved to monty. Please refer to
-monty.json's documentation.
-"""
-
-__author__ = "Shyue Ping Ong"
-__copyright__ = "Copyright 2012, The Materials Project"
-__version__ = "0.1"
-__maintainer__ = "Shyue Ping Ong"
-__email__ = "shyuep@gmail.com"
-__date__ = "Apr 30, 2012"
-
-import json
-import functools
-
-
-def pmg_serialize(method):
-    """
-    Decorator for methods that add MSON serializations keys
-    to the dictionary. See documentation of MSON for more details
-    """
-    @functools.wraps(method)
-    def wrapper(*args, **kwargs):
-        self = args[0]
-        d = method(*args, **kwargs)
-        # Add @module and @class
-        d["@module"] = self.__class__.__module__
-        d["@class"] = self.__class__.__name__
-        return d
-
-    return wrapper
-
-
-def json_pretty_dump(obj, filename):
-    """
-    Serialize obj as a JSON formatted stream to the given filename (
-    pretty printing version)
-    """
-    with open(filename, "wt") as fh:
-        json.dump(obj, fh, indent=4, sort_keys=4)
+from pymatgen.util.serialization import *
