--- pymatgen/serializers/tests/__init__.py
@@ -1,4 +1,0 @@
-"""
-The serializers package implements various tools to encode and decode pymatgen
-objects from commonly used data formats such as json.
-"""
