--- pymatgen/structure_prediction/__init__.py
@@ -1,3 +1,0 @@
-# coding: utf-8
-# Copyright (c) Pymatgen Development Team.
-# Distributed under the terms of the MIT License.
